# packages/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "type": "object"
      },
      "ApiSessionContext": {
        "properties": {
          "ended_at": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "session_id": {
            "type": "string"
          },
          "started_at": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "workspace_id": {
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "session_id",
          "started_at"
        ],
        "type": "object"
      },
...
      "EnqueueMessageResponse": {
        "properties": {
          "accepted_at": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "context": {
            "$ref": "#/components/schemas/ApiSessionContext"
          },
          "run_id": {
            "type": "string"
          }
        },
        "required": [
          "context",
          "run_id",
          "accepted_at"
        ],
        "type": "object"
      },
...
      "ExtensionPart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PartBase"
          },
          {
            "properties": {
              "extensionId": {
                "type": "string"
              },
              "kind": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "name": {
                "type": "string"
              },
              "source": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/FilePartSourceText"
                  }
                ]
              }
            },
            "required": [
              "extensionId",
              "name"
            ],
            "type": "object"
          }
        ]
      },
...
      "FilePart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PartBase"
          },
          {
            "properties": {
              "base64": {
                "type": "string"
              },
              "mediaType": {
                "type": "string"
              },
              "name": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "source": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/FilePartSource"
                  }
                ]
              }
            },
            "required": [
              "base64",
              "mediaType"
            ],
            "type": "object"
          }
        ]
      },
      "FilePartFileSource": {
        "properties": {
          "path": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/FilePartSourceText"
          }
        },
        "required": [
          "text",
          "path"
        ],
        "type": "object"
      },
      "FilePartSource": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FilePartFileSource"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "file"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FilePartSymbolSource"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "symbol"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
      "FilePartSourceText": {
        "properties": {
          "end": {
            "format": "int64",
            "type": "integer"
          },
          "start": {
            "format": "int64",
            "type": "integer"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "value",
          "start",
          "end"
        ],
        "type": "object"
      },
      "FilePartSymbolSource": {
        "properties": {
          "kind": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/FilePartSourceText"
          }
        },
        "required": [
          "text",
          "path",
          "name",
          "kind"
        ],
        "type": "object"
      },
...
      "Message": {
        "properties": {
          "info": {
            "$ref": "#/components/schemas/MessageInfo"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/MessagePart"
            },
            "type": "array"
          }
        },
        "required": [
          "info",
          "parts"
        ],
        "type": "object"
      },
      "MessageInfo": {
        "properties": {
          "completedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "role",
          "createdAt"
        ],
        "type": "object"
      },
      "MessagePart": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/TextPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "text"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReasoningPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "reasoning"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "tool"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExtensionPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "extension"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FilePart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "file"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
...
      "PartBase": {
        "properties": {
          "id": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "messageId"
        ],
        "type": "object"
      },
...
      "ReasoningPart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PartBase"
          },
          {
            "properties": {
              "text": {
                "type": "string"
              }
            },
            "required": [
              "text"
            ],
            "type": "object"
          }
        ]
      },
      "RecordMessageRequest": {
        "properties": {
          "parts": {
            "items": {
              "$ref": "#/components/schemas/SessionCommandInputPart"
            },
            "type": "array"
          }
        },
        "required": [
          "parts"
        ],
        "type": "object"
      },
...
      "SessionCommandExtensionPartInput": {
        "properties": {
          "extensionId": {
            "type": "string"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "source": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FilePartSourceText"
              }
            ]
          }
        },
        "required": [
          "extensionId",
          "name"
        ],
        "type": "object"
      },
      "SessionCommandFilePartInput": {
        "properties": {
          "entryType": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "source": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FilePartSourceText"
              }
            ]
          }
        },
        "required": [
          "path",
          "name"
        ],
        "type": "object"
      },
      "SessionCommandInputPart": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionCommandTextPartInput"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "text"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionCommandExtensionPartInput"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "extension"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionCommandFilePartInput"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "file"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
      "SessionCommandTextPartInput": {
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "type": "object"
      },
...
      "TextPart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PartBase"
          },
          {
            "properties": {
              "text": {
                "type": "string"
              }
            },
            "required": [
              "text"
            ],
            "type": "object"
          }
        ]
      },
      "ToolPart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PartBase"
          },
          {
            "properties": {
              "callId": {
                "type": "string"
              },
              "metadata": {
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                },
                "type": [
                  "object",
                  "null"
                ]
              },
              "state": {
                "$ref": "#/components/schemas/ToolState"
              },
              "tool": {
                "type": "string"
              }
            },
            "required": [
              "callId",
              "tool",
              "state"
            ],
            "type": "object"
          }
        ]
      },
      "ToolState": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolStatePending"
              },
              {
                "properties": {
                  "status": {
                    "enum": [
                      "pending"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolStateRunning"
              },
              {
                "properties": {
                  "status": {
                    "enum": [
                      "running"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolStateCompleted"
              },
              {
                "properties": {
                  "status": {
                    "enum": [
                      "completed"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolStateError"
              },
              {
                "properties": {
                  "status": {
                    "enum": [
                      "error"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
      "ToolStateCompleted": {
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/FilePart"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "input": {
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "metadata": {
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "output": {
            "type": "string"
          },
          "time": {
            "$ref": "#/components/schemas/ToolStateTimeCompleted"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "input",
          "output",
          "title",
          "metadata",
          "time"
        ],
        "type": "object"
      },
      "ToolStateError": {
        "properties": {
          "error": {
            "type": "string"
          },
          "input": {
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "metadata": {
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "time": {
            "$ref": "#/components/schemas/ToolStateTimeRange"
          }
        },
        "required": [
          "input",
          "error",
          "time"
        ],
        "type": "object"
      },
      "ToolStatePending": {
        "properties": {
          "input": {
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "raw": {
            "type": "string"
          }
        },
        "required": [
          "input",
          "raw"
        ],
        "type": "object"
      },
      "ToolStateRunning": {
        "properties": {
          "input": {
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "metadata": {
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "time": {
            "$ref": "#/components/schemas/ToolStateTimeStart"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "input",
          "time"
        ],
        "type": "object"
      },
      "ToolStateTimeCompleted": {
        "properties": {
          "compacted": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "end": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "start": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "type": "object"
      },
      "ToolStateTimeRange": {
        "properties": {
          "end": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "start": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "type": "object"
      },
      "ToolStateTimeStart": {
        "properties": {
          "start": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "start"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-native command bar for macOS",
    "license": {
      "name": ""
    },
    "title": "Cocommand API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/events": {
      "get": {
        "description": "Subscribe to real-time server events via SSE. Optionally filter by session_id.",
        "operationId": "stream_events",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "SSE stream of CoreEvent payloads"
          }
        },
        "tags": [
          "events"
        ]
      }
    },
...
    "/session/command": {
      "get": {
        "description": "Load message history for the current session.",
        "operationId": "session_command_history",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Current session message history"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/command": {
      "post": {
        "description": "Enqueue a message for asynchronous session runtime processing.",
        "operationId": "session_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueMessageResponse"
                }
              }
            },
            "description": "Message accepted and enqueued"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/context": {
      "get": {
        "operationId": "session_context",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSessionContext"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
...
  },
...
}